pyproject.toml
[tool.poetry]
name = "fastapi_poetry"
version = "0.1.0"
description = ""
authors = ["hoangnv0505"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
anyio = "3.6.1"
asgiref = "3.5.2"
bcrypt = "3.2.2"
cffi = "1.15.0"
click = "8.1.3"
colorama = "0.4.4"
config = "0.5.1"
cryptography = "37.0.2"
dependency_injection = "1.2.0"
ecdsa = "0.17.0"
Faker = "13.13.0"
fastapi = "0.78.0"
h11 = "0.13.0"
httptools = "0.4.0"
idna = "3.3"
Jinja2 = "3.1.2"
MarkupSafe = "2.1.1"
passlib = "1.7.4"
pyasn1 = "0.4.8"
pycparser = "2.21"
pydantic = "1.9.1"
PyJWT = "2.4.0"
pyperclip = "1.8.2"
python-dateutil = "2.8.2"
python-decouple = "3.6"
python-dotenv = "0.20.0"
python-jose = "3.3.0"
python-multipart = "0.0.5"
PyYAML = "6.0"
rsa = "4.8"
six = "1.16.0"
sniffio = "1.2.0"
starlette = "0.19.1"
typing-extensions = "4.2.0"
Unidecode = "1.3.4"
uvicorn = "0.17.6"
watchgod = "0.8.2"
websockets = "10.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
